Задание 15.9
