+++ js/form.js
@@ -1,23 +1,29 @@
+import { validateHashtags, getHashtagErrorMessage } from './hashtags.js';
+
+const Pristine = window.Pristine;
+
 const form = document.querySelector('.img-upload__form');
-const fileInput = form.querySelector('#upload-file');
-const overlay = form.querySelector('.img-upload__overlay');
-const cancelButton = form.querySelector('#upload-cancel');
+const fileInput = document.querySelector('.img-upload__input');
+const overlay = document.querySelector('.img-upload__overlay');
+const cancelButton = document.querySelector('.img-upload__cancel');
+const hashtagInput = document.querySelector('.text__hashtags');
+const commentInput = document.querySelector('.text__description');
 const body = document.body;
+const submitButton = document.querySelector('.img-upload__submit');
 
-const hashtagsInput = form.querySelector('.text__hashtags');
-const commentInput = form.querySelector('.text__description');
-const submitButton = form.querySelector('.img-upload__submit');
-
+// Устанавливаем атрибуты формы
 form.action = 'https://29.javascript.htmlacademy.pro/kekstagram';
 form.method = 'POST';
 form.enctype = 'multipart/form-data';
 
+// Открытие формы
 function openForm() {
   overlay.classList.remove('hidden');
   body.classList.add('modal-open');
   document.addEventListener('keydown', onDocumentEscKey);
 }
 
+// Закрытие формы
 function closeForm() {
   overlay.classList.add('hidden');
   body.classList.remove('modal-open');
@@ -26,6 +32,7 @@ function closeForm() {
   document.removeEventListener('keydown', onDocumentEscKey);
 }
 
+// Обработчик Esc
 function onDocumentEscKey(evt) {
   if (evt.key === 'Escape') {
     evt.preventDefault();
@@ -33,7 +40,8 @@ function onDocumentEscKey(evt) {
   }
 }
 
-[hashtagsInput, commentInput].forEach((field) => {
+// Блокируем Esc в полях ввода
+[hashtagInput, commentInput].forEach((field) => {
   field.addEventListener('keydown', (evt) => {
     if (evt.key === 'Escape') {
       evt.stopPropagation();
@@ -41,17 +49,20 @@ function onDocumentEscKey(evt) {
   });
 });
 
+// Открываем форму при выборе файла
 fileInput.addEventListener('change', () => {
   if (fileInput.files.length > 0) {
     openForm();
   }
 });
 
+// Закрываем форму по кнопке
 cancelButton.addEventListener('click', (evt) => {
   evt.preventDefault();
   closeForm();
 });
 
+// Инициализация Pristine
 const pristine = new Pristine(form, {
   classTo: 'img-upload__field-wrapper',
   errorClass: 'img-upload__field-wrapper--invalid',
@@ -59,67 +70,11 @@ const pristine = new Pristine(form, {
   errorTextClass: 'img-upload__error'
 }, true);
 
-const MAX_HASHTAGS = 5;
-const MAX_HASHTAG_LENGTH = 20;
-const HASHTAG_REGEX = /^#[a-zа-яё0-9]+$/i;
-
-function validateHashtags(value) {
-  if (!value.trim()) {
-    return true;
-  }
-
-  const tags = value.trim().split(/\s+/);
-
-  if (tags.length > MAX_HASHTAGS) {
-    return false;
-  }
-
-  const lowerTags = tags.map((t) => t.toLowerCase());
-  const uniqueTags = new Set(lowerTags);
-
-  if (uniqueTags.size !== lowerTags.length) {
-    return false;
-  }
-
-  return tags.every((tag) =>
-    HASHTAG_REGEX.test(tag) &&
-    tag.length <= MAX_HASHTAG_LENGTH &&
-    tag !== '#'
-  );
-}
-
-function hashtagsErrorMessage(value) {
-  const tags = value.trim().split(/\s+/);
-
-  if (tags.length > MAX_HASHTAGS) {
-    return 'Нельзя указать больше пяти хэш-тегов';
-  }
-
-  if (new Set(tags.map((t) => t.toLowerCase())).size !== tags.length) {
-    return 'Хэш-теги не должны повторяться';
-  }
-
-  for (const tag of tags) {
-    if (tag === '#') {
-      return 'Хэш-тег не может состоять из одной решётки';
-    }
-
-    if (!HASHTAG_REGEX.test(tag)) {
-      return 'Неверный формат хэш-тега';
-    }
-
-    if (tag.length > MAX_HASHTAG_LENGTH) {
-      return 'Максимальная длина 20 символов';
-    }
-  }
-
-  return 'Ошибка в хэш-тегах';
-}
-
+// Добавляем валидаторы
 pristine.addValidator(
-  hashtagsInput,
+  hashtagInput,
   validateHashtags,
-  hashtagsErrorMessage
+  getHashtagErrorMessage
 );
 
 pristine.addValidator(
@@ -128,6 +83,7 @@ pristine.addValidator(
   'Комментарий не может быть длиннее 140 символов'
 );
 
+// Обработчик отправки формы
 form.addEventListener('submit', async (evt) => {
   evt.preventDefault();
